Completely grok jQuery text()
The code examples in this book make heavy use of the jQuery text()
method. You need to be aware that the text() method, when used on a
wrapper set containing more than one element, will actually combine and
return a string of text contained in all elements of the wrapper set.
This might be confusing if you were expecting it to return only the text
in the first element of the wrapper set. Below is an example of how the
text() method concatenates the strings found in the elements of a
wrapper set