The concept, behind the concept, behind jQuery
While selecting something and doing something is the core concept behind
jQuery, I would like to extend this concept to include creating
something as well. Therefore, the concept behind jQuery could be
extended to include first creating something new, selecting it, and then
doing something with it. We could call this the concept, behind the
concept, behind jQuery.
What I am trying to make obvious is that you are not stuck with only
selecting something that is already in the DOM. It is additionally
important to grok that jQuery can be used to create new DOM elements and
then do something with these elements.
In the code example below, we create a new <a> element that is not in
the DOM. Once created, it is selected and then manipulated