jQuery requires HTML to run in standards mode or almost-standards mode
There are known issues with jQuery methods not working correctly when a
browser renders an HTML page in quirks mode. Make sure when you are
using jQuery that the browser interprets the HTML in standards mode or
almost standards mode by using a valid doctype.
To ensure proper functionality, code examples in this book use the HTML
5 doctype.