Executing jQuery code when the browser window is completely loaded
Typically, we do not want to wait for the window.onload event. That is
the point of using a custom event like ready() that will execute code
before the window loads, but after the DOM is ready to be traversed and
manipulated.

18
However, sometimes we actually do want to wait. While the custom ready()
event is great for executing code once the DOM is available, we can also
use jQuery to execute code once the entire Web page (including all
assets) is completely loaded.
This can be done by attaching a load event handler to the window object.
jQuery provides the load() event method that can be used to invoke a
function once the window is completely loaded. Below, I provide an
example of the load() event method in use.