Executing jQuery code when DOM is parsed without using ready()
The custom ready() event is not entirely needed. If your JavaScript code
does not affect the DOM, you can include it anywhere in the HTML
document. This means you can avoid the ready() event altogether if your
JavaScript code is not dependent on the DOM being intact.
Most JavaScript nowadays, especially jQuery code, will involve
manipulating the DOM. This means the DOM has to be fully parsed by the
browser in order for you to operate on it. This fact is why developers
have been stuck on the window.onload roller coaster ride for a couple of
years now.
To avoid using the ready() event for code that operates on the DOM, you
can simply place your code in an HTML document before the closing
</body> element. Doing so ensures the DOM is completely loaded, simply
because the browser will parse the document from top to bottom. If you
place your JavaScript code in the document after the DOM elements it
manipulates, there is no need to use the ready() event.
In the example below, I have forgone the use of ready() by simply
placing my script before the document body closes. This is the technique
I use throughout this book and on the majority of sites I build.