Breaking the chain with destructive methods
As mentioned before, not all jQuery methods maintain the chain. Methods
like text() can be chained when used to set the text node of an element.
However, text() actually breaks the chain when used to get the text node
contained within an element.
In the example below, text() is used to set and then get the text
contained within the <p> element.