Using destructive jQuery methods and exiting destruction using end()
jQuery methods that alter the original jQuery wrapper set selected are
considered to be destructive. The reason is that they do not maintain
the original state of the wrapper set. Not to worry; nothing is really
destroyed or removed. Rather, the former wrapper set is attached to a
new set.
However, fun with chaining does not have to stop once the original
wrapped set is altered. Using the end() method, you can back out of any
destructive changes made to the original wrapper set. Examine the usage
of the end() method in the following example to understand how to
operate in and out of DOM elements.