Grokking when the keyword this refers to DOM elements
When attaching events to DOM elements contained in a wrapper set, the
keyword this can be used to refer to the current DOM element invoking
the event. The following example contains jQuery code that will attach a
custom mouseenter event to each <a> element in the page. The native
JavaScript mouseover event fires when the cursor enters or exits a child
element, whereas jQuery's mouseenter does not.