Extracting elements from a wrapper set, using them directly without jQuery
Just because you wrap jQuery functionality around an HTML element does
not mean you lose access to the actual DOM element itself. You can
always extract an element from the wrapper set and operate on the
element via native JavaScript. For example, in the code below I am
setting the title attribute of the <a> element in the HTML page by
manipulating the native title property of the <a> DOM element.