Extracting elements from a wrapper set, using them directly without jQuery
As demonstrated, jQuery provides the handy get() method for accessing
DOM elements at a specific index in the wrapper set.
But there is another option here. You can avoid using the get() method
by simply using the square bracket array notation on the jQuery object
itself. In the context of our prior code example:
This code:
$('a').get(0).title = 'jQuery.com';
Could become this:
$('a')[0].title = 'jQuery.com';
Both allow access to the actual DOM element. Personally, I prefer square
bracket notation. It is faster because it uses native JavaScript to
retrieve the element from an array, instead of passing it to a method.
However, the get() method provides a slick solution for placing all of
the DOM elements into a native array. By simply calling the get() method
without passing it an index parameter, the method will return all of the
DOM elements in the wrapper set in a native JavaScript array.
To demonstrate, let's take get() for a test drive. In the code below, I
am placing all the <a> elements into an array. I then use the array to
access the title property of the third <a> DOM object on the page.