Determining context used by the jQuery function
The default context used by the jQuery function when selecting DOM
elements is the document element (e.g. $('a', document)). This means
that if you do not provide the jQuery function (e.g. jQuery()) with a
second parameter to be used as the context for the DOM query, the
default context used is the document element, more commonly known as
<body>.

32
It is possible to determine the context in which the jQuery function is
performing a DOM query by using the context property. Below I show two
coded examples of retrieving the value of the context property