Grokking the :hidden and :visible filter
The custom jQuery selector filters :hidden and :visible do not take into
account the CSS visibility property as one might expect. The way jQuery
determines if an element is hidden or visible is if the element consumes
any space in the document. To be exact, an element is visible if its
browser-reported offsetWidth or offsetHeight is greater than 0. That
way, an element that might have a CSS display value of block contained
in an element with a display value of none would accurately report that
it is not visible.
Examine the code carefully and make sure you understand why the value
returned is true even though the <div> being selected has an inline
style of display:block.