Using the is() method to return a Boolean value
It is often necessary to determine if the selected set of elements does,
in fact, contain a specific element. Using the is() method, we can check
the current set against an expression/filter. The check will return true
if the set contains at least one element that is selected by the given
expression/filter. If it does not contain the element, a false value is
returned

Notes:  As of jQuery 1.3, the is() method supports all expressions.
Previously, complex expressions such as those containing hierarchy
selectors (such as +, ~, and >) always returned true. Filter is used by
other internal jQuery functions. Therefore, all rules that apply there,
apply here, as well. Some developers use is('.class') to determine if an
element has a specific class. Don't forget that jQuery already has a
method for doing this called hasClass('class'), which can be used on
elements that contain more than one class value. But truth be told,
hasClass() is just a convenient wrapper for the is() method.