Checking wrapper set .length to determine selection
It is possible to determine if your expression has selected anything by
checking if the wrapper set has a length. You can do so by using the
array property length. If the length property does not return 0, then
you know at least one element matches the expression you passed to the
jQuery function. For example, in the code below we check the page for an
element with an id of "notHere." Guess what? It is not there!

Notes:  If it is not obvious, the length property can also report the
number of elements in the wrapper set—stated another way, how many
elements were selected by the expression passed to the jQuery function.