Differences between filtering by numeric order vs. DOM relationships
jQuery provides filters for filtering a wrapper set by an element’s
numerical context within the set.
These filters are:
 :first   :last   :even   :odd   :eq(index)   :gt(index)  
:lt(index)
Notes:  Filters that filter the wrapper set itself do so by filtering
elements in the set at a starting point of 0, or index of 0. For example
:eq(0) and :first access the first element in the
set—$('div:eq(0)')—which is at a 0 index. This is in contrast to the
:nth-child filter that is

39
one-indexed. Meaning, for example, :nth-child(1) will return the first
child element, but trying to use :nth-child(0) will not work. Using
:nth-child(0) will always select nothing.
Using :first will select the first element in the set while :last will
select the last element in the set. Remember that they filter the set
based on the relationship (numerical hierarchy starting at 0) within the
set, but not the elements' relationships in the context of the DOM.
Given this knowledge, it should be obvious why the filters :first,
:last, and :eq(index) will always return a single element.
If it is not obvious, allow me to explain further. The reason that
:first can only return a single element is because there can only be one
element in a set that is considered first when there is only one set.
This should be fairly logical. Examine the code below to see this
concept in action.