Selecting elements by id when the value contains meta-characters
jQuery selectors use a set of meta-characters (e.g. # ~ [] = > ) that
when used as a literal part of a name (e.g. id="#foo[bar]") should be
escaped. It is possible to escape characters by placing two backslashes
before the character. Examine the code below to see how using two
backslashes in the selection expression allows us to select an element
with an id attribute value of #foo[bar].