Stacking selector filters
It is possible to stack selector filters—e.g.
a[title="jQuery"][href^="http://"]. The obvious example of this is
selecting an element that has specific attributes with specific
attribute values. For example, the jQuery code below will only select
<a> elements in the HTML page that:
 Contain an href attribute with a starting value of "http://"   Have a
title attribute with a value of "jQuery"
Only one <a> is being selected